refactor(observe): split observe/pii/types.rs into per-section submodules
Convert `observe/pii/types.rs` (1368 LOC) into `types/` with five
per-section submodules:

- `core.rs` (197 LOC): `name()` + `domain()` methods
- `classification.rs` (~170 LOC): `is_high_risk()`, `is_gdpr_protected()`,
  `is_pci_protected()`, `is_hipaa_protected()`, `is_secret()`
- `mapping.rs` (132 LOC): `From<IdentifierType> for PiiType` impl
- `scan_result.rs` (47 LOC): `PiiScanResult` struct + impl
- `tests.rs` (537 LOC): all `#[cfg(test)]` unit tests

`types/mod.rs` (286 LOC) keeps the `PiiType` enum definition (a single
source of truth for the ~120 variants), the `pub(crate) use scan_result::PiiScanResult`
re-export, and module declarations. The enum body is the bulk — splitting
the variants would mean changing the API; instead the `impl` blocks were
distributed across submodules using Rust's "impls in different files"
pattern (same crate + module, all blocks attach to the same type).

Pure move-only refactor — no logic, signature, visibility, or API
changes. `PiiType` and `PiiScanResult` retain identical external
visibility (`pub` and `pub(crate)` respectively). Follows the precedent
set by #317/#318/#331/#335/#341/#342/#343/#344/#345.

Part of #338 (file-length umbrella). 6 of 7 PRs.
